agent-memory/requirements.txt
pyaudio==0.2.14
wave==0.0.2
python-dotenv==1.0.1
google-generativeai
playsound==1.2.2
pydantic==2.6.1